GEODE-4408: Move hash functions to internal namespace
