+++ cppcache/test/CacheableKeysTest.cpp
@@ -22,40 +22,40 @@
 using namespace apache::geode::client;
 
 TEST(CacheableKeysTest, boolDifferentHashCodes) {
-  EXPECT_NE(serializer::hashcode(false), serializer::hashcode(true))
+  EXPECT_NE(internal::hashcode(false), internal::hashcode(true))
       << "Two different bool values have different hashcodes";
 }
 
 TEST(CacheableKeysTest, int8_tDifferentHashCodes) {
-  EXPECT_NE(serializer::hashcode((int8_t)37), serializer::hashcode((int8_t)42))
+  EXPECT_NE(internal::hashcode((int8_t)37), internal::hashcode((int8_t)42))
       << "Two different int8_t values have different hashcodes";
 }
 
 TEST(CacheableKeysTest, int16_tDifferentHashCodes) {
-  EXPECT_NE(serializer::hashcode((int16_t)37),
-            serializer::hashcode((int16_t)42))
+  EXPECT_NE(internal::hashcode((int16_t)37),
+            internal::hashcode((int16_t)42))
       << "Two different int16_t values have different hashcodes";
 }
 
 TEST(CacheableKeysTest, int32_tDifferentHashCodes) {
-  EXPECT_NE(serializer::hashcode((int32_t)37),
-            serializer::hashcode((int32_t)42))
+  EXPECT_NE(internal::hashcode((int32_t)37),
+            internal::hashcode((int32_t)42))
       << "Two different int32_t values have different hashcodes";
 }
 
 
 TEST(CacheableKeysTest, int64_tDifferentHashCodes) {
-  EXPECT_NE(serializer::hashcode((int64_t)37),
-            serializer::hashcode((int64_t)42))
+  EXPECT_NE(internal::hashcode((int64_t)37),
+            internal::hashcode((int64_t)42))
       << "Two different int64_t values have different hashcodes";
 }
 
 TEST(CacheableKeysTest, floatDifferentHashCodes) {
-  EXPECT_NE(serializer::hashcode(37.F), serializer::hashcode(42.F))
+  EXPECT_NE(internal::hashcode(37.F), internal::hashcode(42.F))
       << "Two different float values have different hashcodes";
 }
 
 TEST(CacheableKeysTest, doubleDifferentHashCodes) {
-  EXPECT_NE(serializer::hashcode(37.), serializer::hashcode(42.))
+  EXPECT_NE(internal::hashcode(37.), internal::hashcode(42.))
       << "Two different double values have different hashcodes";
 }
